ASR-Audio/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai-whisper
numpy==1.24.3
python-multipart==0.0.6
websockets
sounddevice==0.4.6
webrtcvad==2.0.10
